refactor: consolidate invoice print pages into shared PrintableInvoice
